Fix build IGC action
Fix build IGC action used in CI. Every job has been failing for over a week.

The CI got into a bad state, in the newer IGC releases the package names changed
to use dashes instead of underscores, this caused a curl command downloading the packages
to fail, which then got saved to a build cache that's used on every build.

So right now, every build is failing because it's retrieving that invalid build cache
result.

In this PR, I both workaround the build cache issue,
prevent this from happening in the future by erroring in multiple places, and fix the curl script.

The build cache issue can only be worked around because the cache is non-writable, the bad result
would need to be manually deleted by a repo admin. What we can do is just use a new cache key
if the cache result is bad, redownload and store based on the new cache key, and use
the new cache result in the actual IGC build.

I also fixed a CMake issue related to the new shared library name.

I reproduced the exact state of the repo in my fork and verified this fixes it, [here](https://github.com/sarnex/intel-graphics-compiler/actions/runs/12186673537/job/33995744349).

The LLVM 16 build failures are a separate issue.