Emit src file line with RPE
When -emitsrclinetorpe option is passed, VISA attempts to read src line
from debug location. If file/line was found, it's emitted as a comment
for easy reference.

When RPE dump is enabled and -asmOutput is specified, RPE dump is now written
to a file in shader dump with .rpe extension.