Removing C++ code from the C header file kv.h
When included in C code, kv.h causes various compiler errors as a result of
including a C++ header file and using namespace and class names that won't
compile with a C compiler. This patch uses a similar approach to another one
that fixed the same issue in `iga.h`, as well as removes the function
`kv_get_swsb_info` if the compiler is not C++.

This should enable C++ code to continue using the header as if nothing happened,
while also allowing C code to use the header without errors.

Signed-off-by: Ben Olson <matthew.olson@intel.com>