Treat geps with indices zero as bitcast in value tracker
Bitcasting a struct pointer to zero'th field of the struct is essentially
the same as doing a gep with zero as every indice. Because of that the
value tracker becomes confused when a kernel utilizes this property by
using gepping for zero'th elt and bitcasting for them in other places.
This commit fixes the issue by making the value tracker treat such geps
as bitcasts.