Handle loops in liveness analysis for merge allocas
Liveness analysis does not consider the declaration (allocation instruction) as a part of live range of the allocation, instead finds the last possible point that still dominates all uses.
That did not work with loops, where the analysis did not consider the need for the live range to wrap around the loop.
The fix introduced by this change checks if the live range escapes the loop (flows out the exiting block and into the exit block), and if that's true, includes the whole loop in the live range.