Avoid using `%sp` and `%fp` predefined variables
If a kernel uses VLA, but it doesn't use stack calls, then we can initilize
stack-related variables in a limited scope. Only stack and frame pointer need
to be initilized. All other stack-related variables like `ARGV`, `RETV`, etc. don't
need to be initialized as they are not needed for VLA handling.

Also `SP` and `FP` don't need to be initilized as predefined `%sp` and `%fp`
VISA variables. Using predefined variables is only necessary if stack calls
are present in a kernel.

Not using predefined `%sp` and `%fp` VISA variables allows to allocate any
register for stack and frame pointers.