Fix copying initialized members in EmitInsertValueToStruct
Description:
The function getAllDefinedMembers has a bug - it will not return any indices if the insertvalue instruction uses a constant struct instead of undef as a starting point.
For example:
  %29 = insertvalue %__StructSOALayout_ <{ i32 194816, i32 undef, i32 undef, <1 x float> undef }>, i32 %18, 1		; visa id: 45
  %30 = insertvalue %__StructSOALayout_ %29, i32 0, 2		; visa id: 47
In this case, when emitting instructions for %30, we should see that the indices 0 and 1 are initialized (with i32 194816 and i32 %18).
However, we don't and we end up losing those values.

The fix makes getAllDefinedMembers return all initialized indices by recursively traversing the constant struct (recursively in case the struct has structs inside it)