Fix edge case behavior of `lgamma_r`
`lgamma_r` function computes the natural logarithm of the absolute value
of the gamma function. The sign of the gamma function is returned in the
signp argument of `lgamma_r`.

According to the C99 specification, the gamma function should behave as follows:
- If the argument is a negative integer, NaN is returned
- If the argument is ±0, ±∞ is returned

The current implementation in IGC adheres to these specifications but incorrectly
calculates the `signp` argument when the input is ±0.

OpenCL C specification introduces "Additional Requirements Beyond C99":
- lgamma_r(x, signp) returns 0 in signp if x is zero or a negative integer.

Without this change, IGC incorrectly calls the `__spirv_ocl_sign` built-in
function with the result of `gamma(±0)`, which is ±∞. This results in a `signp`
value of ±1, which does not conform to the OpenCL C specification requirement.