rename ShortImplicitPayloadHeader to GlobalOffset
Description:
Compute workloads add following implicit arguments:
 1. payloadHeader - 8 x i32 packing global_id_offset (3 x i32),
    local_size (3 x i32) and 2 x i32 reserved.
 2. enqueued_local_size - 3 x i32

local_size is never used in favour of enqueued_local_size. In the end,
payloadHeader has unused 20 bytes.

shortPayloadHeader was introduced to reduce the size. It incorrectly
used the same name "payloadHeader". To fix the issue, argument must
change name. Instead of "shortPayloadHeader", it is renamed to
"globalOffset", as this is exactly what it is.

---------------------------