Fix crashes encountered during the compilation of Blender kernels utilizing the `SPV_INTEL_bindless_images` extension
The latest version of Blender utilizes the SPV_INTEL_bindless_images
extension in a moreadvanced manner, storing image and sampler handles
in a global variable. This approach has posed a challenging stress test
for the IGC's implementation of the SPV_INTEL_bindless_images extension,
leading to compilation crashes.

These crashes occur due to the use of ValueTracker, which is inappropriate
for the SPV_INTEL_bindless_images extension since images are runtime values
in this context.

Consequently, this change disables value tracking for the purposes of:
- handling SAMPLER_SNAP_WA
- Image3dToImage2darray pass

As a result, both the `SAMPLER_SNAP_WA` and `Image3dToImage2darray` pass
are disabled for kernels utilizing the `SPV_INTEL_bindless_images` extension.