Emit a bitcast after select in SimplifyConstant
In typed pointer mode, LLVM inserts pointer bitcasts between GEPs and
loads of different types. This is not the case in opaque pointer mode
and causes an issue in SimplifyConstant when the GEP + load are replaced
with select of values. This change makes the pass emit a bitcast for
bitwise compatible types.