Don't cast image args to 32-bit integers in built-ins.
This commit updates BiFModule not to truncate the image args to 32-bit
integers. Currently these args are 64-bit truncated to 32-bit integers.
However, because of value tracking, those truncate instructions are
skipped, and the values are treated as 64-bit values anyway. This commit
aims to remove those needless truncate instructions.