Treat madm instruction as a barrier only in local scheduler
To resolve the issue of preventing madm instructions in the math macro interleaved with
other instructions, my previous PR has a defeat which adds the check in G4_INST::isOptBarrier().
That also impacts data flow analysis, RA and SWSB except for local scheduler, which may cause
unexpected performance issue. So, this PR is to revert previous change, and treat madm instruction
as a barrier only in local scheduler by adding the check in CheckBarrier() as this function is
only invoked by local scheduler.