Remove tracking of bindless offset for images
In case of bindless images, NEO runtime passes bindless offset in image
arg of kernel. This bindless offset doesn't need to be tracked because
arg tracking in OCLBIConverter pass is only needed to find image's BTI,
which isn't used with bindless as it's the bindless offset that is used
to find the image's surface state instead of values stored in the
Binding Table.

This commit partially removes the tracking.