Enable `RematAllowExtractElement` by default
When `RematAllowExtractElement` is disabled, the `CloneAddressArithmetic` pass
struggles to accurately evaluate the effects of cloning an address
arithmetic chain if the chain includes any `extractelement` instructions.
The decision on whether to copy the chain is made within the `skipChain`
function. This decision relies on the `FlowMap`, which is incorrectly
updated for chains containing `extractelement` instructions. Consequently,
the pass mistakenly assumes that the address arithmetic chain will merely
be moved down, whereas it is actually copied, leading to suboptimal
optimization.