Ensure correct sequence for `dbg.declare` in `InlineLocalsResolution`
The change ensures correct sequence of `dbg.declare` calls
in function when these are created from global variables.

IGC needs to ensure correct sequence on this step,
and then creates an elf properly. Incorrect sequence of vars
occurs only when we change `DIGlobalVariable` with `isLocal=true` attr
to `DILocalVariable`.

The change also adds the LIT test for this case.