Fix WaveShuffleIndexSinking Pass to handle control flow
Domination tree is used to check if new created instructions for hoisted
variables are placed in dominating basic block and if not - they are moved
to nearest found one. The same mechanism is implemented for merging
function - if "main" WaveShuffleIndex call is not in dominating bb it's
moved to it.