Create CodeScheduling pass
Introduce the CodeScheduling LLVM function pass for local scheduling of the instructions within
a basic block. The goal is to schedule for maximal latency hiding without increasing the register
pressure so much that the spills are introduced.
There is also an option to force scheduling for the minimum possible register pressure.

Added the RegisterPressureTracker helper class to track real register pressure within a basic block,
including proper tracking of "hanging" values, which live intervals are prolonged because of
the regalloc restrictions.

Added the VectorShuffleAnalysis pass to recognize vector shuffle and vector creation patterns
to handle them properly for scheduling purposes.