[LLVM16] Adding support for OpaquePointers in JointMatrixFuncsResolutionPass
This PR adds introduces handling of opaque pointers in LLVM 16.

While for the majority of the JointMatrix code such transition was relatively easy,
then for some specific cases we needed to put more effort.

e.g When argument to the function was provided as an opaque pointer, but it was actually TET/Matrix Type
Then we need to traverse up in order to figure out it's actual type, so we can resolve it correctly.

Such scenario occured when processing access chain.