Refactor enableZEBinary() conditions in OCL
Refactor OCL code paths depending on the enableZEBinary() as if zebin format was always supported and used.

The OpenCLPrintfResolution/basic.ll test was targeting older platforms using the patch token format, so it had to be adjusted for differences in pointer sizes. Specifically, pointers to printf string literals are 64-bit instead of 32-bit.

Some dead code will be left-over after merging, it will be removed in a later PR.