Unused bindless image args treated as bindless fix.
When emitting zeinfo IGC tags addr mode of images with no users as
stateful even if the module is compiled to use bindless images. This
caused NEO to throw an error as it disallows the use of both bindless
and bindful mode in the same module.

This commit sets the default addr mode to bindless for modules that have
UseBindlessImage set to true.