Fix skip to DW_OP_stack_value in `DWARF` split variables
When building a sliced variable, we are emitting two SIMD16 subprograms.
Currently only the upper half lanes (16-31) reaches the `DW_OP_stack_value`
at the end of the source variable's DI Block. The first-half lanes (0-15)
are skipped to the end of the block after emitting the first half,
making the variable unreadable in the debugger.