Skip alloca uses in unreachable BB's when looking for Common Dominator
In Private Memory Resolution, when sinking an alloca instruction, we search for
the nearest common dominator (NCD) among all its uses. If any use resides in an
unreachable basic block, the NCD will not be found, potentially leading to a
segmentation fault. This change ignores  alloca uses in unreachable basic blocks
when searching NCD