PrivateMemoryResolution - fixing cases where elementSize assert was failing
When adding Opaque Pointers support to JointMatrix I've found that 4 test were failing due to this assert:

	info: error, assertion failed: bits == elementSize
	file: Source\IGC\Compiler\Optimizer\OpenCLPasses\PrivateMemory\PrivateMemoryResolution.cpp
	function: TransposeHelperPrivateMem::handleLoadInst
	line: 665

	Failed Tests (4):
	  SYCL :: Matrix/SG32/joint_matrix_bf16_fill_k_cache_unroll.cpp
	  SYCL :: Matrix/SG32/joint_matrix_bf16_fill_k_cache_unroll_init.cpp
	  SYCL :: Matrix/joint_matrix_bf16_fill_k_cache_unroll.cpp
	  SYCL :: Matrix/joint_matrix_bf16_fill_k_cache_unroll_init.cpp

My investigation showed that such resolution path:

alloca -> gep -> load

used invalid vector elements count value, which caused this assert to fail.
To my understanding the reason for this was that we used elementSize saved in "TransposeHelperPrivateMem" instance,
But when we were going thru instructions (alloca->gep->load) then they weren't updated, so there was mismatch.