Fix and improve inline asm
If inline asm's operands are aliases, the current code generate a copy
if the operand is input; and does not handle aliased output operand.

When using copy, it is a little tricky whether to use NoMask or not,
especially for output operands. In addition, using inline asm is most
likely for performance and additional copies should be avoided as much
as possible.

This change fixes output alias operands and also removes copies by
generating visa alias decl with non-zero offset.