Recognize string as printf string, when any of uses is printf (#23337)
Currently constnt string literals used only by printf are recognized
as printf strings and are stored in constant buffer. But when there is
reference to that string, even unused it's no longer recognized as printf
string and it's content is stored in .global buffer. This can lead to
segfault in runtime, as when trying to use printf it will recieve GPU
pointer.

Co-authored-by: Skobejko, Milosz <milosz.skobejko@intel.com>