[LLVM 16] Fixing ResolveOCLRaytracingBuiltins by creating "struct.intel_ray_query_opaque_t" that's not present in generated BiF .bc file
When importing built-in types, the type called "struct.intel_ray_query_opaque_t" was properly imported on typed-pointers mode as:

```
%struct.intel_ray_query_opaque_t = type opaque
```

but on opaque types mode it was not present in the generated BiF .bc file, thus it was not imported.
It caused ResolveOCLRaytracingBuiltins pass to fail because it relied on having  when creating Alloca.

```
auto *allocaType = IGCLLVM::getTypeByName(callInst.getModule(), "struct.intel_ray_query_opaque_t");
auto *alloca = m_builder->CreateAlloca(allocaType);
```

This patch adds workaround for it by creating such type when it is not present.