Workaround for prefetch_cache_control failing due to unknown type size on opaque pointers
Currently `handleCacheControlINTELForPrefetch` requires type size to perform call conversion correctly,
but on opaque pointers such size is not available and we cannot just extract it anyhow.

We're waiting for "OpUntypedPrefetch" extension which will support opaque pointers in such cases.

Meanwhile as of today I'm adding skip of the whole prefetch conversion when opaque pointer type is involved
and we're waiting for update about the status of "OpUntypedPrefetch"