simplify insertelement for large vectors
LLVM's instcombine pass (method
InstCombinerImpl::SimplifyDemandedVectorElts) is able to set input on
first insertelement to poison if all indices in vector are inserted (all
indices are overwritten). This optimization has a hardcoded limit on
vector size in LLVM 15 and 16, which is too short for compute workloads.
Manually optimize such cases in IGC.