Revisit MismatchDetected for LowerGEPForPrivMem
MismatchDetected solved assert in the PrivateMemoryResolution pass
caused by mismatched widths. LowerGEPForPrivMem solves that case
but in case of alloca over the allowed size it will leave mismatched
widths for the later PrivateMemoryResolution causing the assert to
fail. Extended mismatch detection for cases with struct of array/struct
of vector.