[LLVM16][SinkCommonOffsetFromGEP] Detect type mismatch when grouping GEPs
On opaque pointers we could end up with such group:

    %C  = getelementptr inbounds %class.MC_Vector, %class.MC_Vector* %.privateBuffer, i64 0, i32 1
    %C2 = getelementptr inbounds %class.MC_Vector, %class.MC_Vector* %.privateBuffer, i64 0, i32 1
    %A  = getelementptr inbounds %class.MC_Particle, %class.MC_Particle* %.privateBuffer, i64 0, i32 4
    %A2 = getelementptr inbounds %class.MC_Particle, %class.MC_Particle* %.privateBuffer, i64 0, i32 4

The problem is that we have type mismatch here, and later, when such group is being processed it tries
to access indices that aren't available in given type e.g 8th index of 3 element struct.

Such problematic group should be splitted into two different groups and it solves that problem.

    %C  = getelementptr inbounds %class.MC_Vector, %class.MC_Vector* %.privateBuffer, i64 0, i32 1
    %C2 = getelementptr inbounds %class.MC_Vector, %class.MC_Vector* %.privateBuffer, i64 0, i32 1

    %A  = getelementptr inbounds %class.MC_Particle, %class.MC_Particle* %.privateBuffer, i64 0, i32 4
    %A2 = getelementptr inbounds %class.MC_Particle, %class.MC_Particle* %.privateBuffer, i64 0, i32 4

This PR adds check which should result in creation of new group
instead of the merge with existing group that'd result in mismatched types.