Turn invalid asm constraints assert into error
When processing inline vISA, IGC checks that inputs match constraints.
Before this change, if a check failed, the compiler used to silently drop the
instruction or produce an assert, if the compiler was built in debug mode.

After this change, the compiler will throw an error regardless of the build type.