Skip SOA promotion for variable non-promoted type GEPs
Extends the bail out in `SOALayoutChecker::MismatchDetected` to skip not only i8-based
GEPs with non-constant byte indices, but all GEPs with non-constant byte
indices that don't operate on the alloca promoted type.

For example:
```llvm
%8 = alloca [1024 x i8], align 1

memcpy.body:
  %46 = getelementptr <8 x i32>, ptr addrspace(1) %45, i64 %pIV.0
  %47 = getelementptr <8 x i32>, ptr %8, i64 %pIV.0
  %48 = load <8 x i32>, ptr addrspace(1) %46, align 1
  store <8 x i32> %48, ptr %47, align 1
  %49 = add i64 %pIV.0, 1
```
Here the alloca element is `i8`, but element of GEP and store is i32.
On typed pointers, the alloca ptr had to be bitcasted
from *i8 to *i32, and the optimization was disabled because
`SOALayoutChecker::visitBitCastInst` detected the mismatch in size.
This is not possible to check on opaque pointers, so it wasn't detected as
a mismatch, triggering asserts and leading to miscompilations.