Skip split on spill if variable is not live-in/live-out
Split on spill pass splits live-interval of variables that are
live-in/live-out of a loop and are used inside the loop. Splitting such
a spilled variable reduces RA constraints on the split variable, making
it possible to allocate a register for that variable in the loop.

This split must be done only when the variable is live-in to the loop or
live-out of the loop and is defined in the loop. Latter condition is
because in case a variable is written in the loop, it's also spilled to
home location at loop exit.