Fix the algo that generates unique visa names
vISA tries to generate unique names for each variable.
The existing algorithm keeps the name map as small as
possible by not adding visa-generated names into the map.

But the algorithm fails to check if an input name (for
example, llvm name) may collide with visa-generated names.
This leads duplicate visa names.

This change fixes this by checking if an input name may collide
with visa-generated names. This checking is a little expensive.
Thus, to make this checking faster, the suffix added for
visa-generated names is changed to "_#v" from "_#". With this,
a quick check of the last char can tell if it might be visa-generated
names. Only if it ends with 'v' that the expensive checking would
be triggered. As llvm has many names with digits as suffix, this
trick avoid many expensive checking. But the downside is that many
lit tests (around 100) need updating.

The change makes sure any visa-generated new name is unqiue by
checking if it is in the name map. This is also missed in the
existing algorith.