Skip llvm.global.annotations when applying referenced-indirectly attribute
When IGC is checking if function can be accessed indirecty it iterates
over all function users looking for user that is not a function call.

But when pointer to given function is in @llvm.global.annotations it
will also be found and function will be marked as "referenced-indirectly",
despite @llvm.global.annotations being used only for metadata and removed
couple passes later in Dead Global Elimination Pass.