cmake: Add option USE_LINKER to specify linker and prefer lld by default
This change adds USE_LINKER option to cmake that allows for changing the
linker that's going to be used. If not specified, it's gonna probe the
system for a few linkers in some predetermined order and select the
first one that's available. The preference looks like this right now:
`[lld, ld]``, so if you have lld installed on your machine cmake is
automatically going to use it to link for faster build times.
Prior to this change, we just used GNU ld linker which is quite slow.
Note that this option only targets linux systems.