Remove MSAA image function declarations from CTH header
It is possible that, even if IGC requested enabling the cl_khr_gl_msaa_sharing
extension, clang won't enable it because it is not supported on provided OpenCL C
version. `cl_khr_gl_msaa_sharing` is enabled in clang only for OpenCL C 1.2 and
higher. If clang does not enable the extension, the functions declared in the
CTH header for MSAA images will lead to compilation errors, because MSAA types
will not be defined.

These declarations in opencl_cth_released.h are completely redundant, as they are
already declared in clang's built-in headers when the extension is enabled.