Improve i64 emulation optimization for add and sub for VC
When the i64 source values are created as low/high parts combination,
the compiler can eliminate the wrregion and bitcast operations and
propagate the low/high parts directly to the i64 emulation routines.

Also, when the i64 source value of an add or sub instruction is a
zero-extended i32 value, the compiler can optimize the high part
addition or substraction.