Dump assembly when VC kernel name is too long
By default, the Vector Compiler uses the function name as the assembly
file name when dumping the .visaasm and .asm files. However, for some
SYCL/ESIMD workloads, the kernel names can be extremely long and cannot
be used as file names on some OSes (e.g., Linux has a 255-character
limit for file names). This change truncates the kernel names to a
reasonable length (220 characters here, leaving room for prefixes and
suffixes) when generating the assembly file names. If the kernel name
exceeds this length, it is replaced with a generic name "Kernel" plus
a unique kernel ID.